% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{examplePrereg_1}
\alias{examplePrereg_1}
\title{An example (pre)registration specification using the Inclusive
General-Purpose Registration Form}
\format{
An example of a (pre)registration specification
}
\usage{
examplePrereg_1
}
\description{
This is a simple and relatively short partially completed (pre)registration
specification.
}
\keyword{datasets}
