% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_valid_value.R
\name{first_valid_value}
\alias{first_valid_value}
\title{Select the first valid value}
\usage{
first_valid_value(x, selectName = NULL)
}
\arguments{
\item{x}{The vector or list.}

\item{selectName}{Optionally, the name to look at.}
}
\value{
The first valid value (or NULL).
}
\description{
From a vector or list of values, select the first valid value, valid being
defined as a value that is not NULL or NA and has, after being trimmed of
whitespace, a nonzero length. Optionally, only look at the element with a
given name.
}
