% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_almostEmptyForm}
\alias{form_almostEmptyForm}
\title{A mostly empty example form specification}
\format{
A (pre)registration specification
}
\usage{
form_almostEmptyForm
}
\description{
This form specification is mostly empty, so it can be a
useful start if you want to create your own form. The accompanying
Google Sheet, which you can also copy, is
\url{https://docs.google.com/spreadsheets/d/14Qbak7JbBhTqmJaMgJ4tU9ZROaBbUfq37_UzkoHnM60}
}
\keyword{datasets}
