% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_generalPurpose_v1}
\alias{form_generalPurpose_v1}
\title{Inclusive General-Purpose Registration Form}
\format{
A (pre)registration form specification
}
\usage{
form_generalPurpose_v1
}
\description{
This Inclusive General-Purpose Registration Form is intended as a
general-purpose registration form. The form is designed to be applicable
across disciplines (i.e., psychology, economics, law, physics, or any other
field) and across study types (i.e., qualitative studies, quantitative
studies, experiments, systematic reviews, case studies, archive studies,
comparative legal studies, or any other type of study). This form,
therefore, is a fall-back for more specialized forms and can be used if
no specialized form or registration platform is available. If at all
possible, it is recommended to use a specialized form, since this
inclusive general-purpose registration form achieves that inclusiveness
and general-purposeness at the cost of specificity.
}
\keyword{datasets}
