% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_inclDivAddon_v0_1}
\alias{form_inclDivAddon_v0_1}
\alias{inclDivAddon_v0_1}
\title{Inclusivity & Diversity Add-on for preregistration forms}
\format{
A (pre)registration specification
}
\usage{
form_inclDivAddon_v0_1
}
\description{
This form is meant as a generic add-on that can be added to any
preregistration form to ensure that researchers are mindful about
issues related to inclusivity and diversity. With inclusivity we mean
"The fact or quality of being inclusive; (now) esp. the practice or
policy of not excluding any person on the grounds of race, gender,
religion, age, disability, etc.; inclusiveness."
(Oxford University Press, n.d.-b).
}
\details{
With diversity we refer to "The fact, condition, or practice of
including or involving people from a range of different social and
ethnic backgrounds, and (more recently) of different genders, sexual
orientations, etc." (Oxford University Press, n.d.-a).

Oxford University Press. (n.d.-a). Diversity, n. In OED Online. Oxford
University Press. Retrieved November 29, 2022, from
https://www.oed.com/view/Entry/56064

Oxford University Press. (n.d.-b). Inclusivity, n. In OED Online. Oxford
University Press. Retrieved November 29, 2022,
from https://www.oed.com/view/Entry/93584"
}
\keyword{datasets}
