% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_inclSysRev_v0_92}
\alias{form_inclSysRev_v0_92}
\title{Inclusive Systematic Review Registration Form}
\format{
A (pre)registration form specification
}
\source{
\doi{10.31222/osf.io/3nbea}
}
\usage{
form_inclSysRev_v0_92
}
\description{
This Systematic Review Registration Form is intended as a general-purpose
registration form. The form is designed to be applicable to reviews across
disciplines (i.e., psychology, economics, law, physics, or any other field)
and across review types (i.e., scoping review, review of qualitative
studies, meta-analysis, or any other type of review). That means that the
reviewed records may include research reports as well as archive documents,
case law, books, poems, etc. This form, therefore, is a fall-back for more
specialized forms and can be used if no specialized form or registration
platform is available.
}
\keyword{datasets}
