% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_knit.R
\name{form_knit}
\alias{form_knit}
\title{Knit a (pre)registration form into an Rmd file}
\usage{
form_knit(x, section = NULL, headingLevel = 2)
}
\arguments{
\item{x}{The (pre)registration form (as produced by a call
to \code{\link[=form_create]{form_create()}}) or initialized \code{preregr} object
(as produced by a call to \code{\link[=prereg_initialize]{prereg_initialize()}}).}

\item{section}{The section(s) to show; pass \code{NULL} (the default) to show
everything.}

\item{headingLevel}{The level to use for the top-most heading.}
}
\value{
x, invisibly
}
\description{
This function inserts a (pre)registration form, or one or more
sections, into an R Markdown file.
}
\examples{
preregr::form_create(
  title = "Example form",
  version = "0.1.0"
) |>
  preregr::form_knit();
}
