% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_OSFprereg_v1}
\alias{form_OSFprereg_v1}
\title{OSF Prereg form}
\format{
A (pre)registration form specification
}
\source{
\doi{https://doi.org/10.31222/osf.io/epgjd}
}
\usage{
form_OSFprereg_v1
}
\description{
Preregistration is the act of submitting a study plan, ideally also
with analytical plan, to a registry prior to conducting the work.
Preregistration increases the discoverability of research even if it
does not get published further. Adding specific analysis plans can
clarify the distinction between planned, confirmatory tests and
unplanned, exploratory research.
}
\details{
This preprint contains a template for the "OSF Prereg" form available
from the OSF Registry. An earlier version was originally developed for
the Preregistration Challenge, an education campaign designed to
initiate preregistration as a habit prior to data collection in basic
research, funded by the Laura and John Arnold Foundation (now Arnold
Ventures) and conducted by the Center for Open Science. More information
is available at https://www.cos.io/initiatives/prereg/, and other templates
are available at: https://osf.io/zab38/"
}
\keyword{datasets}
