% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_OSFqual1_v1}
\alias{form_OSFqual1_v1}
\title{Qualitative Preregistration Template}
\format{
A (pre)registration specification
}
\usage{
form_OSFqual1_v1
}
\description{
The Qualitative Preregistration Template was created by researchers from
the qualitative community for registration of primarily qualitative work.
It is available as a registration option on the Open Science
Framework (OSF).
}
\details{
This form was added to \code{preregr} by Aleksandra Lazic.
}
\keyword{datasets}
