% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_prereg2D_v1}
\alias{form_prereg2D_v1}
\title{Preregistration Template for Secondary Data Analysis}
\format{
A (pre)registration specification
}
\usage{
form_prereg2D_v1
}
\description{
Please cite the associated paper when using this preregistration
template (see https://doi.org/10.15626/MP.2020.2625).
}
\keyword{datasets}
