% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_that_are_included.R
\docType{data}
\name{form_prpQuant_v1}
\alias{form_prpQuant_v1}
\alias{prpQuant_v1}
\title{Psychological Research Preregistration-Quantitative (aka PRP-QUANT) Template}
\format{
A (pre)registration form specification
}
\source{
\doi{ http://dx.doi.org/10.23668/psycharchives.4584}
}
\usage{
form_prpQuant_v1
}
\description{
As an international effort toward increasing psychology’s commitment to
creating a stronger culture and practice of preregistration, a
multi-society Preregistration Task Force* was formed, following the
2018 meeting of the German Psychological Society in Frankfurt, Germany.
The Task Force created a detailed preregistration template that benefited
from the APA JARS Quantitative Research guidelines, as well as a
comprehensive review of many other preregistration templates.
}
\keyword{datasets}
