% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_to_xlsx.R
\name{form_to_xlsx}
\alias{form_to_xlsx}
\title{Export a (pre)registration form to an Excel spreadsheet}
\usage{
form_to_xlsx(x, file)
}
\arguments{
\item{x}{The (pre)registration form (as produced by a call
to \code{\link[=form_create]{form_create()}}) or initialized \code{preregr} object
(as produced by a call to \code{\link[=prereg_initialize]{prereg_initialize()}}).}

\item{file}{The file to write the spreadsheet to.}
}
\value{
x, invisibly
}
\description{
Export a (pre)registration form to an Excel spreadsheet
}
