% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{forms}
\alias{forms}
\title{Included (pre)registration forms}
\usage{
forms()
}
\value{
Invisibly, a list of form identifiers of the available forms.
}
\description{
Show an overview of all included (pre)registration forms.
}
\examples{
forms();
}
