% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_prereg_spec.R
\name{yaml_to_prereg_spec}
\alias{yaml_to_prereg_spec}
\title{Convert a (pre)registration specification from YAML or JSON}
\usage{
yaml_to_prereg_spec(x)
}
\arguments{
\item{x}{The YAML or JSON as character vector, or a path to a file
containing the YAML or JSON.}
}
\value{
The imported object.
}
\description{
Convert a (pre)registration specification from YAML or JSON
}
\examples{
### Get path to example file
examplePreregFile <-
  system.file(
    "extdata",
    "preregr-spec-example1.yml",
    package = "preregr"
  );

### Load it and show which items are completed
preregr::yaml_to_prereg_spec(
  examplePreregFile
) |>
  preregr::prereg_show_item_completion();
}
