% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg_initialize.R
\name{prereg_initialize}
\alias{prereg_initialize}
\title{Initialize a (pre)registration}
\usage{
prereg_initialize(x, initialText = "Unspecified")
}
\arguments{
\item{x}{The (pre)registration form specification, as a URL to a Google
Sheet or online file or as the path to a locally stored file.}

\item{initialText}{The text to initialize every field with.}
}
\value{
The empty (pre)registration specification.
}
\description{
Initialize a (pre)registration
}
\examples{
preregr::prereg_initialize(
  "inclSysRev_v0_92"
);
}
