% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preregr-package.R
\docType{package}
\name{preregr-package}
\alias{preregr}
\alias{preregr-package}
\title{preregr: Specify (Pre)Registrations and Export Them Human- And Machine-Readably}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Preregistrations, or more generally, registrations, enable explicit timestamped and (often but not necessarily publicly) frozen documentation of plans and expectations as well as decisions and justifications. In research, preregistrations are commonly used to clearly document plans and facilitate justifications of deviations from those plans, as well as decreasing the effects of publication bias by enabling identification of research that was conducted but not published. Like reporting guidelines, (pre)registration forms often have specific structures that facilitate systematic reporting of important items. The 'preregr' package facilitates specifying (pre)registrations in R and exporting them to a human-readable format (using R Markdown partials or exporting to an 'HTML' file) as well as human-readable embedded data (using 'JSON'), as well as importing such exported (pre)registration specifications from such embedded 'JSON'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-packages.gitlab.io/preregr}
  \item Report bugs at \url{https://gitlab.com/r-packages/preregr/-/issues}
}

}
\author{
\strong{Maintainer}: Gjalt-Jorn Peters \email{gjalt-jorn@behaviorchange.eu} (\href{https://orcid.org/0000-0002-0336-9589}{ORCID})

}
\keyword{internal}
