% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialized_data_to_dfs.R
\name{serialized_data_to_dfs}
\alias{serialized_data_to_dfs}
\title{In an object imported from YAML or JSON, convert some elements to dataframes}
\usage{
serialized_data_to_dfs(x)
}
\arguments{
\item{x}{The just imported (pre)registration specification}
}
\value{
The restructured object
}
\description{
In an object imported from YAML or JSON, convert some elements to dataframes
}
