% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg_show_item_content.R
\name{prereg_show_item_content}
\alias{prereg_show_item_content}
\title{Show the specified content for the items in a (pre)registration}
\usage{
prereg_show_item_content(x, section = NULL)
}
\arguments{
\item{x}{The (pre)registration object (as produced by a call to
\code{\link[=prereg_initialize]{prereg_initialize()}}).}

\item{section}{The section(s) to show; pass \code{NULL} (the default) to show
everything.}
}
\value{
x, invisibly
}
\description{
This function shows the specified content for the items in a
(pre)registration, or in one or more sections.
}
\examples{
### Load an example (pre)registration specification
data("examplePrereg_1", package = "preregr");

### Show the item content for the "metadata" section
examplePrereg_1 |>
  preregr::prereg_show_item_content(
    section="metadata"
  );
}
